requirements.txt
alembic==1.4.0
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
astroid==2.3.3
automium==0.2.6
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
Click==7.0
command-not-found==0.3
confusable-homoglyphs==3.2.0
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.0.3
django-bootstrap==0.2.4
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-regex==0.4.1
django-registration==3.1
django-tinymce==2.8.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==20.0.4
httplib2==0.9.2
idna==2.6
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
legacy==0.1.6
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.1.1
markdown2==2.3.8
MarkupSafe==1.0
mccabe==0.6.1
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pbr==5.4.4
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psycopg2==2.8.4
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.1.2
pyperclip==1.7.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-decouple==3.3
python-editor==1.0.4
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
system-service==0.3
systemd-python==234
testresources==2.0.1
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
visitor==0.1.3
wadllib==1.3.2
Werkzeug==0.16.1
whitenoise==5.0.1
wrapt==1.11.2
WTForms==2.2.1
xkit==0.0.0
zope.interface==4.3.2
